洋室の作成結合
