+++ Assets/Scenes/DevScene.unity
@@ -284,7 +284,7 @@ Transform:
   m_ConstrainProportionsScale: 0
   m_Children: []
   m_Father: {fileID: 1227615354}
-  m_RootOrder: 7
+  m_RootOrder: 5
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!114 &643341035
 MonoBehaviour:
@@ -327,7 +327,11 @@ Transform:
   m_LocalPosition: {x: -824.5271, y: 413.97467, z: -16.644642}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
-  m_Children: []
+  m_Children:
+  - {fileID: 1154823183}
+  - {fileID: 1903703919}
+  - {fileID: 1139609501}
+  - {fileID: 1760270684}
   m_Father: {fileID: 1227615354}
   m_RootOrder: 2
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
@@ -348,7 +352,8 @@ MonoBehaviour:
   pa: {fileID: 1838514748}
   cm: {fileID: 1154823184}
   ce: {fileID: 1903703920}
-  cw: {fileID: 1139609502}
+  cwa: {fileID: 1139609502}
+  cws: {fileID: 1760270685}
 --- !u!1 &1103580611
 GameObject:
   m_ObjectHideFlags: 0
@@ -418,12 +423,12 @@ Transform:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 1139609500}
   m_LocalRotation: {x: -0, y: -0, z: -0, w: 1}
-  m_LocalPosition: {x: 3625.8916, y: 724.8197, z: -90.70911}
+  m_LocalPosition: {x: 4450.419, y: 310.84503, z: -74.06447}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children: []
-  m_Father: {fileID: 1227615354}
-  m_RootOrder: 5
+  m_Father: {fileID: 780180543}
+  m_RootOrder: 2
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!114 &1139609502
 MonoBehaviour:
@@ -442,7 +447,8 @@ MonoBehaviour:
   pa: {fileID: 1838514748}
   cm: {fileID: 1154823184}
   ce: {fileID: 1903703920}
-  cw: {fileID: 1139609502}
+  cwa: {fileID: 1139609502}
+  cws: {fileID: 1760270685}
 --- !u!1 &1154823182
 GameObject:
   m_ObjectHideFlags: 0
@@ -468,12 +474,12 @@ Transform:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 1154823182}
   m_LocalRotation: {x: -0, y: -0, z: -0, w: 1}
-  m_LocalPosition: {x: 3625.8916, y: 724.8197, z: -90.70911}
+  m_LocalPosition: {x: 4450.419, y: 310.84503, z: -74.06447}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children: []
-  m_Father: {fileID: 1227615354}
-  m_RootOrder: 3
+  m_Father: {fileID: 780180543}
+  m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!114 &1154823184
 MonoBehaviour:
@@ -492,7 +498,8 @@ MonoBehaviour:
   pa: {fileID: 1838514748}
   cm: {fileID: 1154823184}
   ce: {fileID: 1903703920}
-  cw: {fileID: 1139609502}
+  cwa: {fileID: 1139609502}
+  cws: {fileID: 1760270685}
 --- !u!1 &1227615353
 GameObject:
   m_ObjectHideFlags: 0
@@ -524,14 +531,94 @@ Transform:
   - {fileID: 452271978}
   - {fileID: 1103580613}
   - {fileID: 780180543}
-  - {fileID: 1154823183}
-  - {fileID: 1903703919}
-  - {fileID: 1139609501}
+  - {fileID: 1737145034}
   - {fileID: 1838514749}
   - {fileID: 643341034}
   m_Father: {fileID: 0}
   m_RootOrder: 1
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!1 &1737145033
+GameObject:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  serializedVersion: 6
+  m_Component:
+  - component: {fileID: 1737145034}
+  m_Layer: 0
+  m_Name: EvaluateFloorPlan
+  m_TagString: Untagged
+  m_Icon: {fileID: 0}
+  m_NavMeshLayer: 0
+  m_StaticEditorFlags: 0
+  m_IsActive: 1
+--- !u!4 &1737145034
+Transform:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1737145033}
+  m_LocalRotation: {x: -0, y: -0, z: -0, w: 1}
+  m_LocalPosition: {x: 3625.8916, y: 724.8197, z: -90.70911}
+  m_LocalScale: {x: 1, y: 1, z: 1}
+  m_ConstrainProportionsScale: 0
+  m_Children: []
+  m_Father: {fileID: 1227615354}
+  m_RootOrder: 3
+  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!1 &1760270683
+GameObject:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  serializedVersion: 6
+  m_Component:
+  - component: {fileID: 1760270684}
+  - component: {fileID: 1760270685}
+  m_Layer: 0
+  m_Name: CreateWestern
+  m_TagString: Untagged
+  m_Icon: {fileID: 0}
+  m_NavMeshLayer: 0
+  m_StaticEditorFlags: 0
+  m_IsActive: 1
+--- !u!4 &1760270684
+Transform:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1760270683}
+  m_LocalRotation: {x: -0, y: -0, z: -0, w: 1}
+  m_LocalPosition: {x: 4450.419, y: 310.84503, z: -74.06447}
+  m_LocalScale: {x: 1, y: 1, z: 1}
+  m_ConstrainProportionsScale: 0
+  m_Children: []
+  m_Father: {fileID: 780180543}
+  m_RootOrder: 3
+  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!114 &1760270685
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1760270683}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 67d414d1009e08b41b96f03169d26725, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  lineRenderer: {fileID: 0}
+  pr: {fileID: 1103580612}
+  pa: {fileID: 1838514748}
+  cm: {fileID: 1154823184}
+  ce: {fileID: 1903703920}
+  cwa: {fileID: 1139609502}
+  cws: {fileID: 1760270685}
 --- !u!1 &1838514747
 GameObject:
   m_ObjectHideFlags: 0
@@ -610,7 +697,7 @@ Transform:
   m_ConstrainProportionsScale: 0
   m_Children: []
   m_Father: {fileID: 1227615354}
-  m_RootOrder: 6
+  m_RootOrder: 4
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1 &1903703918
 GameObject:
@@ -637,12 +724,12 @@ Transform:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 1903703918}
   m_LocalRotation: {x: -0, y: -0, z: -0, w: 1}
-  m_LocalPosition: {x: 3625.8916, y: 724.8197, z: -90.70911}
+  m_LocalPosition: {x: 4450.419, y: 310.84503, z: -74.06447}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children: []
-  m_Father: {fileID: 1227615354}
-  m_RootOrder: 4
+  m_Father: {fileID: 780180543}
+  m_RootOrder: 1
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!114 &1903703920
 MonoBehaviour:
@@ -661,4 +748,5 @@ MonoBehaviour:
   pa: {fileID: 1838514748}
   cm: {fileID: 1154823184}
   ce: {fileID: 1903703920}
-  cw: {fileID: 1139609502}
+  cwa: {fileID: 1139609502}
+  cws: {fileID: 1760270685}
