+++ Assets/Scripts/CreateWestern.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 67d414d1009e08b41b96f03169d26725
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
